# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Gestion Magasin Documentation",
        "description": "Implementation solution de Gestion de Magasin avec Laravel",
        "version": "1.0.0"
    },
    "paths": {
...
        "/facturefournisseurs": {
            "get": {
                "tags": [
                    "Affiche  Facture de Fournisseurs"
                ],
                "summary": "Affiche la liste des Facture de Fournisseurs",
                "description": "Retourner Touts les Facture de Fournisseurs",
                "operationId": "Affiche Facture de Fournisseurs",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Ajouter Facture de Fournisseur"
                ],
                "summary": "Ajoute Facture de Fournisseurs",
                "description": "Ajouter  Facture de Fournisseur et Retourne les Valeurs Ajouter",
                "operationId": "Store Facture de Fournisseur",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
        "/facturefournisseurs/{facturefournisseur}": {
            "get": {
                "tags": [
                    "Affiche Facture de Fournisseur par ID"
                ],
                "summary": "Affiche le Facture de Fournisseur par ID",
                "description": "Retourner le Facture de Fournisseur",
                "operationId": "Show Facture de Fournisseur",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Mise en Jour Facture et Lignefacture"
                ],
                "summary": "Modifier Facture et Lignefacture",
                "description": "Retourner Facture et Lignefacture Modifier",
                "operationId": "Modifier Facture et Lignefacture",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Supprimer Facture et Lignefacture"
                ],
                "summary": "Supprimer un Facture et Lignefacture",
                "description": "Supprimer un Facture et Lignefacture",
                "operationId": "Delete Facture et Lignefacture",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {}
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "not found"
                    }
                }
            }
        },
...
    }
}